feat(frontend): adapt /devices/chat page for mobile devices
## Changes

**Component Architecture (Code Reuse):**
- Split `/devices/chat/page.tsx` into three components following responsive pattern:
  - `page.tsx` - Router with mobile/desktop detection and dynamic imports
  - `DeviceChatPageDesktop.tsx` - Desktop implementation (≥768px)
  - `DeviceChatPageMobile.tsx` - Mobile implementation (≤767px)
- Reuse all business logic between components (device selection, task handling, WebSocket)
- Only styling and layout differ between implementations

**Mobile-Specific Features:**
- Slide-out drawer sidebar via TaskSidebar's MobileSidebar component
- Touch-friendly device selector with 44px minimum touch targets
- Compact Select component replacing native select for better mobile UX
- Beta badge in TopNavigation header (amber style matching /devices page)
- ThemeToggle instead of GithubStarButton on mobile
- Mobile-optimized spacing and layouts

**Desktop Features Preserved:**
- Resizable sidebar with collapse support
- Full navigation and toolbar
- All existing functionality unchanged

**Technical Details:**
- Uses `useIsMobile()` hook for responsive breakpoint detection (≤767px)
- Dynamic imports with `ssr: false` for code splitting
- URL parameter sync (TaskParamSync, DeviceTaskSync) maintained
- Proper handling of offline devices with disabled state
- TypeScript strict mode compliance
- English code comments following project standards

## Testing
- Mobile sidebar drawer functionality
- Device selector on mobile (Select component)
- Beta badge display in TopNavigation
- Touch target sizes (44px minimum)
- Desktop functionality unchanged
- URL parameter syncing
- Offline device handling
